+++ projects/storefrontapp/src/app/app.module.ts
@@ -13,9 +13,9 @@ import {
   JsonLdBuilderModule,
   StorefrontComponent,
 } from '@spartacus/storefront';
-import { b2bFeature } from '../environments/b2c/b2b.environment';
-import { b2cFeature } from '../environments/b2c/b2c.environment';
-import { cdsFeature } from '../environments/cds/cds.environment';
+import { b2bFeature } from '../environments/b2c/b2b.feature';
+import { b2cFeature } from '../environments/b2c/b2c.feature';
+import { cdsFeature } from '../environments/cds/cds.feature';
 import { environment } from '../environments/environment';
 import { TestOutletModule } from '../test-outlets/test-outlet.module';
