--- projects/storefrontapp/src/environments/b2c/b2b.environment.ts
@@ -1,5 +1,0 @@
-import { FeatureEnvironment } from '../models/feature.environment.model';
-
-export const b2bFeature: FeatureEnvironment = {
-  imports: [],
-};
