+++ projects/storefrontapp/src/environments/b2c/b2b.feature.ts
@@ -1,0 +1,5 @@
+import { FeatureEnvironment } from '../models/feature.model';
+
+export const b2bFeature: FeatureEnvironment = {
+  imports: [],
+};
