--- projects/storefrontapp/src/environments/cds/cds.environment.ts
@@ -1,5 +1,0 @@
-import { FeatureEnvironment } from '../models/feature.environment.model';
-
-export const cdsFeature: FeatureEnvironment = {
-  imports: [],
-};
