+++ projects/storefrontapp/src/environments/cds/cds.feature.ts
@@ -1,0 +1,5 @@
+import { FeatureEnvironment } from '../models/feature.model';
+
+export const cdsFeature: FeatureEnvironment = {
+  imports: [],
+};
